+++ example/lib/pages/button_view.dart
@@ -1,3 +1,5 @@
+import 'package:example/markdown/markdown_value.dart';
+import 'package:example/markdown/markdown_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_miui/flutter_miui.dart';
 
@@ -10,7 +12,24 @@ class _ButtonViewState extends State<ButtonView> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: MIUIAppBar(title: '按钮'),
+      appBar: MIUIAppBar(
+        title: Text('按钮 Demo'),
+        actions: <Widget>[
+          IconButton(
+            icon: Icon(Icons.code),
+            onPressed: () {
+              Navigator.of(context).push(
+                MaterialPageRoute(
+                  builder: (context) => MarkdownView(
+                    data: MV.buttonAll,
+                    title: 'Button Code',
+                  ),
+                ),
+              );
+            },
+          ),
+        ],
+      ),
       body: ListView(
         physics: MIUIScrollPhysics(),
         children: <Widget>[