+++ lib/flutter_miui.dart
@@ -7,3 +7,4 @@ export 'src/miui_appbar.dart';
 export 'src/miui_button.dart';
 export 'src/miui_menu.dart';
 export 'src/miui_global.dart';
+export 'src/miui_textfield.dart';