+++ lib/src/miui_appbar.dart
@@ -1,23 +1,14 @@
 import 'package:flutter/material.dart';
 
-class MIUIAppBar extends StatefulWidget implements PreferredSizeWidget {
-  MIUIAppBar({Key key, @required this.title, this.actions}) : super(key: key);
-  final String title;
-  final List<Widget> actions;
-  @override
-  State<StatefulWidget> createState() => _MIUIAppBarState();
-
-  @override
-  Size get preferredSize => Size.fromHeight(60);
-}
-
-class _MIUIAppBarState extends State<MIUIAppBar> {
-  @override
-  Widget build(BuildContext context) {
-    return AppBar(
-      title: Text(widget.title),
-      centerTitle: true,
-      actions: widget.actions,
-    );
-  }
+class MIUIAppBar extends AppBar {
+  MIUIAppBar({
+    Widget title,
+    List<Widget> actions,
+    Widget leading,
+  }) : super(
+          leading: leading,
+          title: title,
+          actions: actions,
+          centerTitle: true,
+        );
 }